requirements.txt
alembic==1.8.1
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dominate==2.7.0
exceptiongroup==1.0.3
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.44
tomli==2.0.1
urllib3==1.26.12
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1